feat: Update nano-build options and presets
This commit updates the nano-build options and presets in the nano-build.cjs file. It introduces a new developmentOptions object for sourcemap and sourcesContent settings. The dropLabels option is used to remove the __dev_mode__ label. The getOptions function now checks for the preset name and merges the corresponding presetOptions. The options object is then updated based on the devMode value, using either the nano-build-development or nano-build-production settings. Finally, null fields are removed from the esbuildOptions.